refactor/frontend: tried to use leaflet instead of vue-leaflet
