+++ frontend/.browserslistrc
@@ -1,0 +1,4 @@
+> 1%
+last 2 versions
+not dead
+not ie 11
